refactor: add group folder
