+++ src/app/(content)/about/_components/Slider.tsx
