+++ src/app/(content)/welcome/_components/AboutSection.tsx
