+++ src/app/(content)/welcome/_components/FaqSection.tsx
