+++ src/app/(content)/welcome/_components/Hero.tsx
