+++ src/app/(content)/welcome/_components/JoinSection.tsx
