+++ src/app/(content)/welcome/_components/TeamsSection.tsx
