+++ src/app/(content)/welcome/page.tsx
@@ -1,11 +1,11 @@
 import type { Metadata } from "next";
-import Header from "./_components/Header";
+import Header from "../../../components/Header";
 import Hero from "./_components/Hero";
 import AboutSection from "./_components/AboutSection";
 import TeamsSection from "./_components/TeamsSection";
 import FaqSection from "./_components/FaqSection";
 import JoinSection from "./_components/JoinSection";
-import Footer from "./_components/Footer";
+import Footer from "../../../components/Footer";
 
 export const metadata: Metadata = {
   title: "デジクリの紹介 - 芝浦工業大学 デジクリ",
@@ -39,18 +39,12 @@ export const metadata: Metadata = {
 
 export default function WelcomePage() {
   return (
-    <>
-      <Header />
-      <div className="body flex min-h-screen flex-col lg:ml-[25%]">
-        <main className="flex-grow">
-          <Hero />
-          <AboutSection />
-          <TeamsSection />
-          <FaqSection />
-          <JoinSection />
-        </main>
-        <Footer />
-      </div>
-    </>
+    <main className="flex-grow">
+      <Hero />
+      <AboutSection />
+      <TeamsSection />
+      <FaqSection />
+      <JoinSection />
+    </main>
   );
 }
