--- src/app/about/layout.tsx
@@ -1,7 +1,0 @@
-export default function AboutLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <>{children}</>;
-}
